+++ buildpacks/yarn/CHANGELOG.md
@@ -4,6 +4,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # [Unreleased]
 - Add license to buildpack.toml ([#17](https://github.com/heroku/buildpacks-node/pull/17))
+- Flush cache when stack image changes ([#28](https://github.com/heroku/buildpacks-node/pull/28))
 
 ## [0.1.1] 2021/01/20